+++ package.json
@@ -11,6 +11,7 @@
         "prop-types": "^15.8.1",
         "react": "^18.2.0",
         "react-dom": "^18.2.0",
+        "react-router-dom": "^5.3.4",
         "react-scripts": "5.0.1",
         "web-vitals": "^2.1.4"
     },
