+++ src/app/App.js
@@ -1,8 +1,25 @@
 import React from "react";
-import Products from "./components/products";
+import { Route, Switch, Redirect } from "react-router-dom";
+
+import Products from "./layouts/products";
+import Login from "./layouts/login";
+import Main from "./layouts/main";
+import NavBar from "./components/navBar";
+// import ProductPage from "./components/productPage";
 
 function App() {
-    return <Products />;
+    return (
+        <div>
+            <NavBar />
+            <Switch>
+                {/* <Route path="/products/:prodId?" component={ProductPage} /> */}
+                <Route path="/products/:prodId?" component={Products} />
+                <Route path="/login" component={Login} />
+                <Route path="/" exact component={Main} />
+                <Redirect to="/" />
+            </Switch>
+        </div>
+    );
 }
 
 export default App;
