+++ src/app/components/navBar.jsx
@@ -1,0 +1,26 @@
+import React from "react";
+import { Link } from "react-router-dom";
+
+const NavBar = () => {
+    return (
+        <ul className="nav">
+            <li className="nav-item">
+                <Link className="nav-link" aria-current="page" to="/">
+                    Main
+                </Link>
+            </li>
+            <li className="nav-item">
+                <Link className="nav-link" aria-current="page" to="/products">
+                    Products
+                </Link>
+            </li>
+            <li className="nav-item">
+                <Link className="nav-link" aria-current="page" to="/login">
+                    Login
+                </Link>
+            </li>
+        </ul>
+    );
+};
+
+export default NavBar;
